+++ next-shopping/src/@types/cart.d.ts
@@ -1,0 +1,8 @@
+import { ItemResultsType } from "./firebase";
+
+export type CartItemType = ItemResultsType & { quantity: number };
+
+export type CartType = {
+  items: CartItemType[];
+  total: number;
+};
