+++ next-shopping/src/@types/login.d.ts
@@ -1,0 +1,4 @@
+export type LoginStatusType = {
+  isLoggedIn: boolean;
+  uid?: string
+}
